+++ lib/muser_priv.h
@@ -38,8 +38,8 @@
 extern char *irq_to_str[];
 
 int
-muser_pci_hdr_access(lm_ctx_t *lm_ctx, size_t *count,
-                     loff_t *pos, bool write, char *buf);
+muser_pci_hdr_access(lm_ctx_t *lm_ctx, uint32_t *count,
+                     uint64_t *pos, bool write, char *buf);
 
 lm_reg_info_t *
 lm_get_region_info(lm_ctx_t *lm_ctx);
@@ -85,6 +85,72 @@ send_recv_vfio_user_msg(int sock, uint16_t msg_id, enum vfio_user_command cmd,
                         int *send_fds, size_t fd_count,
                         struct vfio_user_header *hdr,
                         void *recv_data, size_t recv_len);
+
+/* FIXME copied from include/linux/stddef.h, is this OK license-wise? */
+#define sizeof_field(TYPE, MEMBER) sizeof((((TYPE *)0)->MEMBER))
+#define offsetofend(TYPE, MEMBER) \
+       (offsetof(TYPE, MEMBER) + sizeof_field(TYPE, MEMBER))
+
+static inline ssize_t get_minsz(unsigned int cmd)
+{
+	switch (cmd) {
+	case VFIO_DEVICE_GET_INFO:
+		return offsetofend(struct vfio_device_info, num_irqs);
+	case VFIO_DEVICE_GET_REGION_INFO:
+		return offsetofend(struct vfio_region_info, offset);
+	case VFIO_DEVICE_GET_IRQ_INFO:
+		return offsetofend(struct vfio_irq_info, count);
+	case VFIO_DEVICE_SET_IRQS:
+		return offsetofend(struct vfio_irq_set, count);
+	case VFIO_GROUP_GET_STATUS:
+		return offsetofend(struct vfio_group_status, flags);
+	case VFIO_GET_API_VERSION:
+		return 0;
+	case VFIO_CHECK_EXTENSION:
+	case VFIO_GROUP_SET_CONTAINER:
+	case VFIO_GROUP_UNSET_CONTAINER:
+	case VFIO_SET_IOMMU:
+		return sizeof(int);
+	case VFIO_IOMMU_GET_INFO:
+		return offsetofend(struct vfio_iommu_type1_info, iova_pgsizes);
+	case VFIO_IOMMU_MAP_DMA:
+		return offsetofend(struct vfio_iommu_type1_dma_map, size);
+	case VFIO_IOMMU_UNMAP_DMA:
+		return offsetofend(struct vfio_iommu_type1_dma_unmap, size);
+	case VFIO_GROUP_GET_DEVICE_FD:
+	case VFIO_DEVICE_RESET:
+		return 0;
+	}
+	return -EOPNOTSUPP;
+}
+
+static inline const char* vfio_cmd_to_str(int cmd) {
+        switch (cmd) {
+                case VFIO_GET_API_VERSION: return "VFIO_GET_API_VERSION";
+                case VFIO_CHECK_EXTENSION: return "VFIO_CHECK_EXTENSION";
+                case VFIO_SET_IOMMU: return "VFIO_SET_IOMMU";
+                case VFIO_GROUP_GET_STATUS: return "VFIO_GROUP_GET_STATUS";
+                case VFIO_GROUP_SET_CONTAINER: return "VFIO_GROUP_SET_CONTAINER";
+                case VFIO_GROUP_UNSET_CONTAINER: return "VFIO_GROUP_UNSET_CONTAINER";
+                case VFIO_GROUP_GET_DEVICE_FD: return "VFIO_GROUP_GET_DEVICE_FD";
+                case VFIO_DEVICE_GET_INFO: return "VFIO_DEVICE_GET_INFO";
+                case VFIO_DEVICE_GET_REGION_INFO: return "VFIO_DEVICE_GET_REGION_INFO";
+                case VFIO_DEVICE_GET_IRQ_INFO: return "VFIO_DEVICE_GET_IRQ_INFO";
+                case VFIO_DEVICE_SET_IRQS: return "VFIO_DEVICE_SET_IRQS";
+                case VFIO_DEVICE_RESET: return "VFIO_DEVICE_RESET";
+                case VFIO_IOMMU_GET_INFO: return "VFIO_IOMMU_GET_INFO/VFIO_DEVICE_GET_PCI_HOT_RESET_INFO/VFIO_IOMMU_SPAPR_TCE_GET_INFO";
+                case VFIO_IOMMU_MAP_DMA: return "VFIO_IOMMU_MAP_DMA/VFIO_DEVICE_PCI_HOT_RESET";
+                case VFIO_IOMMU_UNMAP_DMA: return "VFIO_IOMMU_UNMAP_DMA";
+                case VFIO_IOMMU_ENABLE: return "VFIO_IOMMU_ENABLE";
+                case VFIO_IOMMU_DISABLE: return "VFIO_IOMMU_DISABLE";
+                case VFIO_EEH_PE_OP: return "VFIO_EEH_PE_OP";
+                case VFIO_IOMMU_SPAPR_REGISTER_MEMORY: return "VFIO_IOMMU_SPAPR_REGISTER_MEMORY";
+                case VFIO_IOMMU_SPAPR_UNREGISTER_MEMORY: return "VFIO_IOMMU_SPAPR_UNREGISTER_MEMORY";
+                case VFIO_IOMMU_SPAPR_TCE_CREATE: return "VFIO_IOMMU_SPAPR_TCE_CREATE";
+                case VFIO_IOMMU_SPAPR_TCE_REMOVE: return "VFIO_IOMMU_SPAPR_TCE_REMOVE";
+        }
+        return NULL;
+}
 
 #endif /* MUSER_PRIV_H */
